Move <GeneralContentControls> into its own file
